Colorization and Line Segment
Added colorized electric field arrows. Blue is small magnitude and red is large magnitude. The arrows are also scaled with magnitude so the view isn't cluttered with arrows getting in the way of seeing the important stuff.
Also added functionality in the shader for computing the electric field of line segments.